GestureManager.cs moved hand related interactions into HandsManager.cs
